refactor(porygon-typeorm): update models folder
