+++ packages/porygon-typeorm/src/persistence/builders/buildPorygonTypeOrmSourceModuleMetadataAsync.spec.ts
@@ -9,7 +9,7 @@ import {
 } from '@cuaklabs/iocuak';
 import { DataSourceOptions } from 'typeorm';
 
-import { PorygonTypeOrmSourceModuleSymbolsMap } from '../models/typeorm/PorygonTypeOrmSourceModuleSymbolsMap';
+import { PorygonTypeOrmSourceModuleSymbolsMap } from '../models/domain/PorygonTypeOrmSourceModuleSymbolsMap';
 import { buildPorygonTypeOrmSourceModule } from './buildPorygonTypeOrmSourceModule';
 import { buildPorygonTypeOrmSourceModuleMetadataAsync } from './buildPorygonTypeOrmSourceModuleMetadataAsync';
