+++ packages/porygon-typeorm/src/persistence/builders/buildPorygonTypeOrmSourceModuleSymbolsMap.ts
@@ -1,4 +1,4 @@
-import { PorygonTypeOrmSourceModuleSymbolsMap } from '../models/typeorm/PorygonTypeOrmSourceModuleSymbolsMap';
+import { PorygonTypeOrmSourceModuleSymbolsMap } from '../models/domain/PorygonTypeOrmSourceModuleSymbolsMap';
 
 const VERSION: string = 'v1';
